Code de-duplication for element filtering using composable.
